requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
pillow==11.3.0
psycopg2==2.9.10
pytils==0.4.4
sqlparse==0.5.3
tzdata==2025.2
